backend/infrastructure/image_pipeline/requirements.txt
Pillow
boto3
pillow-heif
requests